summarization-service/requirements.txt
Flask
python-dotenv
langchain
langchain-aws
boto3
